+++ src/test_definition.rs
@@ -5,7 +5,7 @@ use crate::test_file::{
 };
 use chrono::{offset::TimeZone, Days, Local, Months, NaiveDate};
 use hyper::Method;
-use log::trace;
+use log::{debug, error, trace};
 use serde::{Deserialize, Serialize};
 use std::cell::Cell;
 use std::collections::HashSet;
@@ -47,14 +47,18 @@ pub struct HttpParameter {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RequestBody {
+    pub data: serde_json::Value,
+    matches_variable: Cell<bool>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct RequestDescriptor {
     pub method: HttpVerb,
     pub url: String,
     pub params: Vec<HttpParameter>,
     pub headers: Vec<HttpHeader>,
-    pub body: Option<serde_json::Value>,
-
-    body_matches_variable: Cell<bool>,
+    pub body: Option<RequestBody>,
 }
 
 // TODO: add validation logic to verify the descriptor is valid
@@ -84,14 +88,30 @@ impl RequestDescriptor {
             None => Vec::new(),
         };
 
+        let request_body = match request.body {
+            Some(b) => Some(RequestBody {
+                data: b,
+                matches_variable: Cell::from(false),
+            }),
+            None => None,
+        };
+
         Ok(RequestDescriptor {
             method: request.method.unwrap_or(HttpVerb::Get),
             url: request.url,
             params: validated_params,
             headers: validated_headers,
-            body: request.body,
-            body_matches_variable: Cell::from(false),
+            body: request_body,
         })
+    }
+
+    pub fn new_opt(
+        request_opt: Option<UnvalidatedRequest>,
+    ) -> Result<Option<RequestDescriptor>, ValidationError> {
+        match request_opt {
+            Some(request) => Ok(Some(RequestDescriptor::new(request)?)),
+            None => Ok(None),
+        }
     }
 
     pub fn validate(&self) -> bool {
@@ -109,9 +129,7 @@ pub struct CompareDescriptor {
     pub headers: Vec<HttpHeader>,
     pub add_headers: Vec<HttpHeader>,
     pub ignore_headers: Vec<String>,
-    pub body: Option<serde_json::Value>,
-
-    body_matches_variable: Cell<bool>,
+    pub body: Option<RequestBody>,
 }
 
 impl CompareDescriptor {
@@ -188,6 +206,14 @@ impl CompareDescriptor {
                     };
                 }
 
+                let compare_body = match request.body {
+                    Some(b) => Some(RequestBody {
+                        data: b,
+                        matches_variable: Cell::from(false),
+                    }),
+                    None => None,
+                };
+
                 Ok(Some(CompareDescriptor {
                     method: request.method.unwrap_or(HttpVerb::Get),
                     url: request.url,
@@ -197,8 +223,7 @@ impl CompareDescriptor {
                     headers: validated_headers,
                     add_headers: validated_add_headers,
                     ignore_headers: validated_ignore_headers,
-                    body: request.body,
-                    body_matches_variable: Cell::from(false),
+                    body: compare_body,
                 }))
             }
             None => Ok(None),
@@ -216,11 +241,20 @@ pub struct ResponseExtraction {
     pub field: String,
 }
 
+impl ResponseExtraction {
+    pub fn new() -> ResponseExtraction {
+        ResponseExtraction {
+            name: "".to_string(),
+            field: "".to_string(),
+        }
+    }
+}
+
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ResponseDescriptor {
     pub status: Option<u16>,
     pub headers: Vec<HttpHeader>,
-    pub body: Option<serde_json::Value>,
+    pub body: Option<RequestBody>,
     pub ignore: Vec<String>,
     pub extract: Vec<ResponseExtraction>,
 }
@@ -231,8 +265,8 @@ impl ResponseDescriptor {
         response: Option<UnvalidatedResponse>,
     ) -> Result<Option<ResponseDescriptor>, ValidationError> {
         match response {
-            Some(req) => {
-                let validated_headers = match req.headers {
+            Some(res) => {
+                let validated_headers = match res.headers {
                     Some(headers) => headers
                         .iter()
                         .map(|h| HttpHeader {
@@ -244,20 +278,28 @@ impl ResponseDescriptor {
                     None => Vec::new(),
                 };
 
-                let validated_ignore = match req.ignore {
+                let validated_ignore = match res.ignore {
                     Some(ignore) => ignore,
                     None => Vec::new(),
                 };
 
-                let validated_extraction = match req.extract {
+                let validated_extraction = match res.extract {
                     Some(extract) => extract,
                     None => Vec::new(),
                 };
 
+                let response_body = match res.body {
+                    Some(b) => Some(RequestBody {
+                        data: b,
+                        matches_variable: Cell::from(false),
+                    }),
+                    None => None,
+                };
+
                 Ok(Some(ResponseDescriptor {
-                    status: req.status,
+                    status: res.status,
                     headers: validated_headers,
-                    body: req.body,
+                    body: response_body,
                     ignore: validated_ignore,
                     extract: validated_extraction,
                 }))
@@ -290,6 +332,16 @@ pub struct Modifier {
     pub operation: String,
     pub value: String,
     pub unit: String,
+}
+
+impl Modifier {
+    pub fn new() -> Modifier {
+        Modifier {
+            operation: "".to_string(),
+            value: "".to_string(),
+            unit: "".to_string(),
+        }
+    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -346,7 +398,7 @@ impl TestVariable {
             VariableTypes::Datetime => String::from(""),
         };
 
-        trace!("generate_value result: {}", result);
+        debug!("generate_value result: {}", result);
 
         return result;
     }
@@ -354,11 +406,11 @@ impl TestVariable {
     fn generate_int_value(&self, iteration: u32) -> String {
         match &self.value {
             serde_yaml::Value::Number(v) => {
-                trace!("number expression: {:?}", v);
+                debug!("number expression: {:?}", v);
                 return format!("{}", v);
             }
             serde_yaml::Value::Sequence(seq) => {
-                trace!("sequence expression: {:?}", seq);
+                debug!("sequence expression: {:?}", seq);
                 let test = &seq[iteration as usize];
                 let test_string = match test {
                     serde_yaml::Value::Number(st) => st.as_i64().unwrap_or(0),
@@ -367,7 +419,7 @@ impl TestVariable {
                 return format!("{}", test_string);
             }
             serde_yaml::Value::Mapping(map) => {
-                trace!("map expression: {:?}", map);
+                debug!("map expression: {:?}", map);
                 return String::from("");
             }
             _ => {
@@ -379,7 +431,7 @@ impl TestVariable {
     fn generate_string_value(&self, iteration: u32, global_variables: Vec<TestVariable>) -> String {
         match &self.value {
             serde_yaml::Value::String(v) => {
-                trace!("string expression: {:?}", v);
+                debug!("string expression: {:?}", v);
 
                 if v.contains("$") {
                     let mut modified_value = v.clone();
@@ -400,7 +452,7 @@ impl TestVariable {
                 return format!("{}", v);
             }
             serde_yaml::Value::Sequence(seq) => {
-                trace!("sequence expression: {:?}", seq);
+                debug!("sequence expression: {:?}", seq);
                 let test = &seq[iteration as usize];
                 let test_string = match test {
                     serde_yaml::Value::String(st) => st.to_string(),
@@ -426,7 +478,7 @@ impl TestVariable {
                 return format!("{}", test_string);
             }
             serde_yaml::Value::Mapping(map) => {
-                trace!("map expression: {:?}", map);
+                debug!("map expression: {:?}", map);
                 return String::from("");
             }
             _ => {
@@ -439,7 +491,7 @@ impl TestVariable {
         // TODO: Add proper error handling
         match &self.value {
             serde_yaml::Value::String(v) => {
-                trace!("string expression: {:?}", v);
+                debug!("string expression: {:?}", v);
                 let mut result_date;
 
                 let modified_value = if v.contains("$") {
@@ -513,7 +565,7 @@ impl TestVariable {
                 return format!("{}", result_date.format("%Y-%m-%d"));
             }
             serde_yaml::Value::Sequence(seq) => {
-                trace!("sequence expression: {:?}", seq);
+                debug!("sequence expression: {:?}", seq);
                 let test = &seq[iteration as usize];
 
                 let test_string: &str = match test {
@@ -552,14 +604,14 @@ impl TestVariable {
                         );
                     }
                     Err(e) => {
-                        println!("parse_attempt failed");
-                        println!("{}", e);
+                        error!("parse_attempt failed");
+                        error!("{}", e);
                         return String::from("");
                     }
                 }
             }
             serde_yaml::Value::Mapping(map) => {
-                trace!("map expression: {:?}", map);
+                debug!("map expression: {:?}", map);
                 return String::from("");
             }
             _ => {
@@ -656,22 +708,26 @@ impl RequestResponseDescriptor {
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct CleanupDescriptor {
-    pub onsuccess: Option<RequestResponseDescriptor>,
-    pub onfailure: Option<RequestResponseDescriptor>,
-    pub request: RequestDescriptor,
+    pub onsuccess: Option<RequestDescriptor>,
+    pub onfailure: Option<RequestDescriptor>,
+    pub request: Option<RequestDescriptor>,
 }
 
 impl CleanupDescriptor {
-    pub fn new_opt(
+    pub fn new(
         cleanup_opt: Option<UnvalidatedCleanup>,
-    ) -> Result<Option<CleanupDescriptor>, ValidationError> {
+    ) -> Result<CleanupDescriptor, ValidationError> {
         match cleanup_opt {
-            Some(cleanup) => Ok(Some(CleanupDescriptor {
-                onsuccess: RequestResponseDescriptor::new_opt(cleanup.onsuccess)?,
-                onfailure: RequestResponseDescriptor::new_opt(cleanup.onfailure)?,
-                request: RequestDescriptor::new(cleanup.request)?,
-            })),
-            None => Ok(None),
+            Some(cleanup) => Ok(CleanupDescriptor {
+                onsuccess: RequestDescriptor::new_opt(cleanup.onsuccess)?,
+                onfailure: RequestDescriptor::new_opt(cleanup.onfailure)?,
+                request: RequestDescriptor::new_opt(cleanup.request)?,
+            }),
+            None => Ok(CleanupDescriptor {
+                onsuccess: None,
+                onfailure: None,
+                request: None,
+            }),
         }
     }
 
@@ -684,6 +740,7 @@ impl CleanupDescriptor {
 pub struct TestDefinition {
     pub name: Option<String>,
     pub id: String,
+    pub environment: Option<String>,
     pub requires: Option<String>,
     pub tags: Vec<String>,
     pub iterate: u32,
@@ -691,7 +748,7 @@ pub struct TestDefinition {
     pub global_variables: Vec<TestVariable>,
     pub stages: Vec<StageDescriptor>,
     pub setup: Option<RequestResponseDescriptor>,
-    pub cleanup: Option<CleanupDescriptor>,
+    pub cleanup: CleanupDescriptor,
 }
 
 // TODO: add validation logic to verify the descriptor is valid
@@ -715,6 +772,7 @@ impl TestDefinition {
         let td = TestDefinition {
             name: test.name,
             id: test.id.unwrap_or(generated_id).to_lowercase(),
+            environment: test.env,
             requires: test.requires,
             tags: new_tags,
             iterate: test.iterate.unwrap_or(1),
@@ -727,98 +785,141 @@ impl TestDefinition {
                 test.stages,
             )?,
             setup: RequestResponseDescriptor::new_opt(test.setup)?,
-            cleanup: CleanupDescriptor::new_opt(test.cleanup)?,
+            cleanup: CleanupDescriptor::new(test.cleanup)?,
         };
 
         td.update_variable_matching();
         Ok(td)
     }
 
+    fn update_request_variables(request: &RequestDescriptor, var_pattern: &str) {
+        for header in request.headers.iter() {
+            if header.matches_variable.get() {
+                continue;
+            }
+
+            if header.value.contains(var_pattern) {
+                header.matches_variable.set(true);
+                debug!("setting match true: {}", header.header);
+            }
+        }
+
+        for param in request.params.iter() {
+            if param.matches_variable.get() {
+                continue;
+            }
+
+            if param.value.contains(var_pattern) {
+                param.matches_variable.set(true);
+                debug!("setting match true: {}", param.param);
+            }
+        }
+
+        if let Some(body) = request.body.as_ref() {
+            let body_data = match serde_json::to_string(&body.data) {
+                Ok(s) => Some(s),
+                Err(_) => None,
+            };
+
+            if let Some(b) = &body_data {
+                if b.contains(var_pattern) {
+                    body.matches_variable.set(true);
+                    debug!("request body match true: {}", var_pattern);
+                }
+            }
+        }
+    }
+
+    fn update_compare_variables(compare: &CompareDescriptor, var_pattern: &str) {
+        for header in compare.headers.iter() {
+            if header.matches_variable.get() {
+                continue;
+            }
+
+            if header.value.contains(var_pattern) {
+                header.matches_variable.set(true);
+                debug!("setting match true: {}", header.header);
+            }
+        }
+
+        for param in compare.params.iter() {
+            if param.matches_variable.get() {
+                continue;
+            }
+
+            if param.value.contains(var_pattern) {
+                param.matches_variable.set(true);
+                debug!("setting match true: {}", param.param);
+            }
+        }
+
+        if let Some(body) = &compare.body {
+            let body_data = match serde_json::to_string(&body.data) {
+                Ok(s) => Some(s),
+                Err(_) => None,
+            };
+
+            if let Some(b) = &body_data {
+                if b.contains(var_pattern) {
+                    body.matches_variable.set(true);
+                    debug!("compare body match true: {}", var_pattern);
+                }
+            }
+        }
+    }
+
+    fn update_response_variables(response: &ResponseDescriptor, var_pattern: &str) {
+        for header in response.headers.iter() {
+            if header.matches_variable.get() {
+                continue;
+            }
+
+            if header.value.contains(var_pattern) {
+                header.matches_variable.set(true);
+                debug!("setting match true: {}", header.header);
+            }
+        }
+
+        if let Some(body) = response.body.as_ref() {
+            let body_data = match serde_json::to_string(&body.data) {
+                Ok(s) => Some(s),
+                Err(_) => None,
+            };
+
+            if let Some(b) = &body_data {
+                if b.contains(var_pattern) {
+                    body.matches_variable.set(true);
+                    debug!("response body match true: {}", var_pattern);
+                }
+            }
+        }
+    }
+
     fn update_variable_matching(&self) {
-        trace!("updating variable matching");
-        let mut body_str: Option<String> = None;
-        let mut compare_body_str: Option<String> = None;
+        trace!("scanning test definition for variable pattern matches");
 
         for variable in self.variables.iter().chain(self.global_variables.iter()) {
             let var_pattern = format!("${}$", variable.name.trim());
+            // debug!("pattern: {}", var_pattern);
 
             if let Some(setup) = self.setup.as_ref() {
-                for header in setup.request.headers.iter() {
-                    if header.value.contains(var_pattern.as_str()) {
-                        header.matches_variable.set(true);
-                        trace!("setting match true: {}", header.header);
-                    }
-                }
-
-                for param in setup.request.params.iter() {
-                    if param.value.contains(var_pattern.as_str()) {
-                        param.matches_variable.set(true);
-                        trace!("setting match true: {}", param.param);
-                    }
-                }
-
-                if body_str.is_none() && !setup.request.body.is_none() {
-                    body_str = match &setup.request.body {
-                        Some(v) => match serde_json::to_string(&v) {
-                            Ok(s) => Some(s),
-                            Err(_) => None,
-                        },
-                        None => None,
-                    };
-                }
-
-                if !setup.request.body_matches_variable.get() {
-                    if let Some(body) = &body_str {
-                        if body.contains(var_pattern.as_str()) {
-                            setup.request.body_matches_variable.set(true);
-                            trace!("request body match true: {}", var_pattern.as_str());
-                        }
-                    }
-                }
+                TestDefinition::update_request_variables(&setup.request, var_pattern.as_str());
 
                 if let Some(response) = &setup.response {
-                    for header in response.headers.iter() {
-                        if header.value.contains(var_pattern.as_str()) {
-                            header.matches_variable.set(true);
-                            trace!("response setting match true: {}", header.header);
-                        }
-                    }
+                    TestDefinition::update_response_variables(&response, var_pattern.as_str());
                 }
             }
 
-            if let Some(cleanup) = self.cleanup.as_ref() {
-                for header in cleanup.request.headers.iter() {
-                    if header.value.contains(var_pattern.as_str()) {
-                        header.matches_variable.set(true);
-                        trace!("setting match true: {}", header.header);
-                    }
-                }
+            if let Some(request) = &self.cleanup.request {
+                TestDefinition::update_request_variables(&request, var_pattern.as_str());
+            }
 
-                for param in cleanup.request.params.iter() {
-                    if param.value.contains(var_pattern.as_str()) {
-                        param.matches_variable.set(true);
-                        trace!("setting match true: {}", param.param);
-                    }
-                }
+            if let Some(onsuccess) = &self.cleanup.onsuccess {
+                TestDefinition::update_request_variables(&onsuccess, var_pattern.as_str());
+            }
 
-                if body_str.is_none() && !cleanup.request.body.is_none() {
-                    body_str = match &cleanup.request.body {
-                        Some(v) => match serde_json::to_string(&v) {
-                            Ok(s) => Some(s),
-                            Err(_) => None,
-                        },
-                        None => None,
-                    };
-                }
-
-                if !cleanup.request.body_matches_variable.get() {
-                    if let Some(body) = &body_str {
-                        if body.contains(var_pattern.as_str()) {
-                            cleanup.request.body_matches_variable.set(true);
-                            trace!("request body match true: {}", var_pattern.as_str());
-                        }
-                    }
-                }
+            if let Some(onfailure) = &self.cleanup.onfailure {
+                TestDefinition::update_request_variables(&onfailure, var_pattern.as_str());
             }
         }
 
@@ -829,102 +930,22 @@ impl TestDefinition {
                 .chain(self.variables.iter().chain(self.global_variables.iter()))
             {
                 let var_pattern = format!("${}$", variable.name.trim());
-                trace!("pattern: {}", var_pattern);
-                for header in stage.request.headers.iter() {
-                    if header.value.contains(var_pattern.as_str()) {
-                        header.matches_variable.set(true);
-                        trace!("setting match true: {}", header.header);
-                    }
-                }
+                // debug!("pattern: {}", var_pattern);
 
-                for param in stage.request.params.iter() {
-                    if param.value.contains(var_pattern.as_str()) {
-                        param.matches_variable.set(true);
-                        trace!("setting match true: {}", param.param);
-                    }
-                }
-
-                if body_str.is_none() && !stage.request.body.is_none() {
-                    body_str = match &stage.request.body {
-                        Some(v) => match serde_json::to_string(&v) {
-                            Ok(s) => Some(s),
-                            Err(_) => None,
-                        },
-                        None => None,
-                    };
-                }
-
-                if !stage.request.body_matches_variable.get() {
-                    if let Some(body) = &body_str {
-                        if body.contains(var_pattern.as_str()) {
-                            stage.request.body_matches_variable.set(true);
-                            trace!("request body match true: {}", var_pattern.as_str());
-                        }
-                    }
-                }
+                TestDefinition::update_request_variables(&stage.request, var_pattern.as_str());
 
                 if let Some(compare) = &stage.compare {
-                    for header in compare.headers.iter() {
-                        if header.value.contains(var_pattern.as_str()) {
-                            header.matches_variable.set(true);
-                            trace!("compare setting match true: {}", header.header);
-                        }
-                    }
-
-                    for header in compare.add_headers.iter() {
-                        if header.value.contains(var_pattern.as_str()) {
-                            header.matches_variable.set(true);
-                            trace!("compare add_header setting match true: {}", header.header);
-                        }
-                    }
-
-                    for param in compare.params.iter() {
-                        if param.value.contains(var_pattern.as_str()) {
-                            param.matches_variable.set(true);
-                            trace!("compare setting match true: {}", param.param);
-                        }
-                    }
-
-                    for param in compare.add_params.iter() {
-                        if param.value.contains(var_pattern.as_str()) {
-                            param.matches_variable.set(true);
-                            trace!("compare add_param setting match true: {}", param.param);
-                        }
-                    }
-
-                    if compare_body_str.is_none() && !compare.body.is_none() {
-                        compare_body_str = match &compare.body {
-                            Some(v) => match serde_json::to_string(&v) {
-                                Ok(s) => Some(s),
-                                Err(_) => None,
-                            },
-                            None => None,
-                        };
-                    }
-
-                    if !compare.body_matches_variable.get() {
-                        if let Some(body) = &compare_body_str {
-                            if body.contains(var_pattern.as_str()) {
-                                compare.body_matches_variable.set(true);
-                                trace!("compare body match true: {}", var_pattern.as_str());
-                            }
-                        }
-                    }
+                    TestDefinition::update_compare_variables(&compare, var_pattern.as_str());
                 }
 
                 if let Some(response) = &stage.response {
-                    for header in response.headers.iter() {
-                        if header.value.contains(var_pattern.as_str()) {
-                            header.matches_variable.set(true);
-                            trace!("response setting match true: {}", header.header);
-                        }
-                    }
+                    TestDefinition::update_response_variables(&response, var_pattern.as_str());
                 }
             }
         }
     }
 
-    fn get_url(
+    pub fn get_url(
         &self,
         iteration: u32,
         url: &str,
@@ -971,62 +992,7 @@ impl TestDefinition {
         }
     }
 
-    pub fn get_setup_request_url(&self, iteration: u32) -> String {
-        match self.setup.as_ref() {
-            Some(setup) => self.get_url(
-                iteration,
-                &setup.request.url,
-                &setup.request.params,
-                &self.variables,
-            ),
-            None => String::from(""),
-        }
-    }
-
-    pub fn get_cleanup_request_url(&self, iteration: u32) -> String {
-        match self.cleanup.as_ref() {
-            Some(cleanup) => self.get_url(
-                iteration,
-                &cleanup.request.url,
-                &cleanup.request.params,
-                &self.variables,
-            ),
-            None => String::from(""),
-        }
-    }
-
-    pub fn get_stage_request_url(&self, stage_index: usize, iteration: u32) -> String {
-        let stage = self.stages.get(stage_index).unwrap();
-        self.get_url(
-            iteration,
-            &stage.request.url,
-            &stage.request.params,
-            &[&stage.variables[..], &self.variables[..]].concat(),
-        )
-    }
-
-    pub fn get_stage_compare_url(&self, stage_index: usize, iteration: u32) -> String {
-        let stage = self.stages.get(stage_index).unwrap();
-        match stage.compare.as_ref() {
-            Some(compare) => {
-                let params = if compare.params.len() > 0 {
-                    &compare.params
-                } else {
-                    &stage.request.params
-                };
-                self.get_url(
-                    iteration,
-                    &compare.url,
-                    params,
-                    &[&stage.variables[..], &self.variables[..]].concat(),
-                )
-            }
-            None => String::from(""),
-        }
-    }
-
     fn get_processed_param(&self, parameter: &HttpParameter, iteration: u32) -> (String, String) {
-        // println!("processing param: {}", parameter.param);
         for variable in self.variables.iter().chain(self.global_variables.iter()) {
             let var_pattern = format!("${}$", variable.name);
 
@@ -1085,35 +1051,8 @@ impl TestDefinition {
         }
     }
 
-    pub fn get_cleanup_request_headers(&self, iteration: u32) -> Vec<(String, String)> {
-        match self.cleanup.as_ref() {
-            Some(cleanup) => cleanup
-                .request
-                .headers
-                .iter()
-                .map(|h| {
-                    if h.matches_variable.get() {
-                        let header = self.get_processed_header(h, iteration);
-                        (header.0, header.1)
-                    } else {
-                        (h.header.clone(), h.value.clone())
-                    }
-                })
-                .collect(),
-            None => Vec::new(),
-        }
-    }
-
-    pub fn get_stage_request_headers(
-        &self,
-        stage_index: usize,
-        iteration: u32,
-    ) -> Vec<(String, String)> {
-        self.stages
-            .get(stage_index)
-            .unwrap()
-            .request
-            .headers
+    pub fn get_headers(&self, headers: &Vec<HttpHeader>, iteration: u32) -> Vec<(String, String)> {
+        headers
             .iter()
             .map(|h| {
                 if h.matches_variable.get() {
@@ -1124,6 +1063,13 @@ impl TestDefinition {
                 }
             })
             .collect()
+    }
+
+    pub fn get_cleanup_request_headers(&self, iteration: u32) -> Vec<(String, String)> {
+        match &self.cleanup.request {
+            Some(request) => self.get_headers(&request.headers, iteration),
+            None => Vec::new(),
+        }
     }
 
     pub fn get_stage_compare_headers(
@@ -1176,108 +1122,87 @@ impl TestDefinition {
         }
     }
 
-    fn get_body(
+    pub fn get_body(
         &self,
-        iteration: u32,
-        matches: bool,
-        body: &Option<serde_json::Value>,
+        request: &RequestDescriptor,
         variables: &Vec<TestVariable>,
+        iteration: u32,
     ) -> Option<serde_json::Value> {
-        if !matches || body.is_none() {
-            return body.clone();
-        }
-
-        let mut body_str = match body {
-            Some(v) => match serde_json::to_string(&v) {
-                Ok(s) => s,
-                Err(_) => "".to_string(),
-            },
-            None => "".to_string(),
-        };
-
-        for variable in variables.iter().chain(self.global_variables.iter()) {
-            let var_pattern = format!("${}$", variable.name);
-
-            if !body_str.contains(var_pattern.as_str()) {
-                continue;
+        if let Some(body) = &request.body {
+            if !body.matches_variable.get() {
+                return Some(body.data.clone());
             }
 
-            let replacement = variable.generate_value(iteration, self.global_variables.clone());
-            body_str = body_str.replace(var_pattern.as_str(), replacement.as_str());
+            let mut body_str = match serde_json::to_string(&body.data) {
+                Ok(s) => s,
+                Err(_) => "".to_string(),
+            };
+
+            for variable in variables.iter().chain(self.global_variables.iter()) {
+                let var_pattern = format!("${}$", variable.name);
+
+                if !body_str.contains(var_pattern.as_str()) {
+                    continue;
+                }
+
+                let replacement = variable.generate_value(iteration, self.global_variables.clone());
+                body_str = body_str.replace(var_pattern.as_str(), replacement.as_str());
+            }
+
+            return match serde_json::from_str(body_str.as_str()) {
+                Ok(result) => Some(result),
+                Err(_) => None,
+            };
         }
 
-        match serde_json::from_str(body_str.as_str()) {
-            Ok(result) => Some(result),
-            Err(_) => None,
-        }
+        None
     }
 
-    pub fn get_setup_request_body(&self, iteration: u32) -> Option<serde_json::Value> {
-        match self.setup.as_ref() {
-            Some(setup) => self.get_body(
-                iteration,
-                setup.request.body_matches_variable.get(),
-                &setup.request.body,
-                &self.variables,
-            ),
-            None => None,
-        }
-    }
-
-    pub fn get_cleanup_request_body(&self, iteration: u32) -> Option<serde_json::Value> {
-        match self.cleanup.as_ref() {
-            Some(cleanup) => self.get_body(
-                iteration,
-                cleanup.request.body_matches_variable.get(),
-                &cleanup.request.body,
-                &self.variables,
-            ),
-            None => None,
-        }
-    }
-
-    pub fn get_stage_request_body(
+    pub fn get_compare_body(
         &self,
-        stage_index: usize,
+        compare: &CompareDescriptor,
+        variables: &Vec<TestVariable>,
         iteration: u32,
     ) -> Option<serde_json::Value> {
-        let stage = self.stages.get(stage_index).unwrap();
-        self.get_body(
-            iteration,
-            stage.request.body_matches_variable.get(),
-            &stage.request.body,
-            &[&stage.variables[..], &self.variables[..]].concat(),
-        )
-    }
+        if let Some(body) = &compare.body {
+            if !body.matches_variable.get() {
+                return Some(body.data.clone());
+            }
 
-    pub fn get_stage_compare_body(
-        &self,
-        stage_index: usize,
-        iteration: u32,
-    ) -> Option<serde_json::Value> {
-        let stage = self.stages.get(stage_index).unwrap();
+            let mut body_str = match serde_json::to_string(&body.data) {
+                Ok(s) => s,
+                Err(_) => "".to_string(),
+            };
 
-        match stage.compare.as_ref() {
-            Some(compare) => self.get_body(
-                iteration,
-                compare.body_matches_variable.get(),
-                &compare.body,
-                &[&stage.variables[..], &self.variables[..]].concat(),
-            ),
-            None => None,
+            for variable in variables.iter().chain(self.global_variables.iter()) {
+                let var_pattern = format!("${}$", variable.name);
+
+                if !body_str.contains(var_pattern.as_str()) {
+                    continue;
+                }
+
+                let replacement = variable.generate_value(iteration, self.global_variables.clone());
+                body_str = body_str.replace(var_pattern.as_str(), replacement.as_str());
+            }
+
+            return match serde_json::from_str(body_str.as_str()) {
+                Ok(result) => Some(result),
+                Err(_) => None,
+            };
         }
+
+        None
     }
 
     pub fn validate(&self) -> bool {
+        trace!("validating test definition");
         let mut valid_td = true;
 
         if let Some(setup) = &self.setup {
             valid_td &= setup.validate();
         }
 
-        if let Some(cleanup) = &self.cleanup {
-            valid_td &= cleanup.validate();
-        }
+        valid_td &= self.cleanup.validate();
 
         for stage in self.stages.iter() {
             valid_td &= stage.request.validate();